restructure project, fix unfortunate issue with line breaks
