+++ CMakeLists.txt
@@ -1,17 +1,15 @@
-cmake_minimum_required(VERSION 3.0)
+cmake_minimum_required(VERSION 3.5)
+project(jet VERSION 0.1.0 LANGUAGES C)
 
+# enable testing through cmake
+enable_testing()
+
+# dependencies
 set(CURSES_NEED_NCURSES TRUE)
 find_package(Curses REQUIRED)
 
-set(SOURCE_FILES src/term/jet.c
-                 src/core/jet.h
-                 src/core/buffer.c
-                 src/core/buffer.h
-                 src/core/file.c
-                 src/core/file.h
-                 src/core/line.c
-                 src/core/line.h)
+# core module
+add_subdirectory(core)
 
-add_executable(jet ${SOURCE_FILES})
-
-target_link_libraries(jet PRIVATE ncurses)
+add_executable(jet term/src/jet.c)
+target_link_libraries(jet PRIVATE core_lib ncurses)
