+++ core/include/core/jet.h
@@ -4,8 +4,8 @@
  * Copyright (c) 2018 Ethan Martin
  */
 
-#include "buffer.h"
-#include "file.h"
+#include <core/buffer.h>
+#include <core/file.h>
 
 /* fatal error, print a message and terminate the program */
 void die(const char *error, int code);
